building and linting
