+++ components/ui/AlertDialog.tsx
@@ -10,8 +10,6 @@ import { DialogTitle, DialogDescription } from '~/components/ui/dialog';
 
 const AlertDialog = AlertDialogPrimitive.Root;
 
-const AlertDialogTrigger = AlertDialogPrimitive.Trigger;
-
 const AlertDialogPortal = ({
   ...props
 }: AlertDialogPrimitive.AlertDialogPortalProps) => (
