+++ components/ui/TimeAgo.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect, useMemo } from 'react';
+import { dateOptions } from '~/fresco.config';
 import { withNoSSRWrapper } from '~/utils/NoSSRWrapper';
-import { dateOptions } from './FormattedDate';
 
 type TimeAgoProps = {
   date: Date | string | number;
