+++ components/ui/typography/Heading.tsx
@@ -4,7 +4,6 @@ import { cva, type VariantProps } from 'class-variance-authority';
 import React from 'react';
 import { cn } from '~/utils/shadcn';
 import { Slot } from '@radix-ui/react-slot';
-import { motion } from 'framer-motion';
 
 export const headingVariants = cva('text-balance', {
   variants: {
@@ -57,7 +56,5 @@ const Heading = React.forwardRef<HTMLElement, HeadingProps>(
 );
 
 Heading.displayName = 'Heading';
-
-const MotionHeading = motion(Heading);
 
 export default Heading;
