+++ lib/interviewer/containers/SlidesForm/SlideFormEdge.js
@@ -92,7 +92,5 @@ const withStore = connect((state, props) => {
   };
 });
 
-const EnhancedSlideFormEdge = withEdgeProps(SlideFormEdge);
-
 
 export default compose(withStore, withEdgeProps)(SlideFormEdge);
