+++ lib/interviewer/containers/SlidesForm/SlideFormNode.js
@@ -69,7 +69,5 @@ const withNodeProps = withProps(({ item }) => ({
   initialValues: item?.[entityAttributesProperty],
 }));
 
-const EnhancedSlideFormNode = withNodeProps(SlideFormNode);
-
 
 export default withNodeProps(SlideFormNode);
