--- lib/interviewer/ducks/modules/errors.js
@@ -1,13 +1,0 @@
-import { actionCreators as dialogActions } from './dialogs';
-
-const withErrorDialog =
-  (actionCreator) =>
-  (...args) =>
-  (dispatch) => {
-    dispatch({ type: 'PLAY_SOUND', sound: 'error' });
-    const action = actionCreator(...args);
-    dispatch(action);
-    dispatch(dialogActions.openDialog({ type: 'Error', error: action.error }));
-  };
-
-export { withErrorDialog };
