+++ lib/interviewer/ducks/modules/network.js
@@ -8,11 +8,6 @@ import { SET_SERVER_SESSION } from './setServerSession';
 /*
  * For actionCreators see `src/ducks/modules/sessions`
  */
-
-// Property names passed to user worker scripts
-const primaryKeyPropertyForWorker = 'networkCanvasId';
-const nodeTypePropertyForWorker = 'networkCanvasType';
-
 const ADD_NODE = 'ADD_NODE';
 const ADD_NODE_TO_PROMPT = 'ADD_NODE_TO_PROMPT';
 const BATCH_ADD_NODES = 'BATCH_ADD_NODES';
