+++ lib/interviewer/ducks/modules/reset.js
@@ -4,8 +4,6 @@ import { actionCreators as deviceActions } from './deviceSettings';
 import { get } from '../../utils/lodash-replacements';
 
 const RESET_STATE = 'RESET_STATE';
-const RESET_EDGES_OF_TYPE = 'RESET/EDGES_OF_TYPE';
-const RESET_PROPERTY_FOR_ALL_NODES = 'RESET/PROPERTY_FOR_ALL_NODES';
 
 const resetPropertyForAllNodes = (property) => (dispatch, getState) => {
   const { activeSessionId } = getState();
