+++ lib/interviewer/ducks/modules/ui.js
@@ -41,10 +41,6 @@ const actionCreators = {
   toggle,
 };
 
-const actionTypes = {
-  UPDATE,
-  TOGGLE,
-};
 
 export {
   actionCreators,
