+++ lib/interviewer/selectors/interface.js
@@ -28,33 +28,6 @@ export const getNodeVariables = createSelector(
   },
 );
 
-const makeGetNodeVariables = () => getNodeVariables;
-
-const getEdgeVariables = createSelector(
-  getProtocolCodebook,
-  getSubjectType,
-  (codebook, edgeType) => {
-    const edgeInfo = codebook.edge;
-    return edgeInfo?.[edgeType]?.variables || {};
-  },
-);
-
-const makeGetEdgeVariables = () => getEdgeVariables;
-
-const getVariableOptions = createSelector(
-  (state, props) => {
-    if (props?.subject?.entity === 'edge') {
-      return getEdgeVariables(state, props);
-    }
-
-    return getNodeVariables(state, props);
-  },
-  (state, { variable }) => variable,
-  (variables, variable) => {
-    return variables[variable]?.options || []
-  },
-);
-
 export const makeGetVariableOptions = (includeOtherVariable = false) =>
   createSelector(
     getNodeVariables,
@@ -112,8 +85,6 @@ export const getNetworkEntitiesForType = createSelector(
     return [network.ego];
   },
 );
-
-const makeNetworkEntitiesForType = () => getNetworkEntitiesForType;
 
 /**
  * makeNetworkNodesForType()
@@ -173,8 +144,6 @@ export const getNetworkNodesForPrompt = createSelector(
   (nodes, promptId) => filter(nodes, (node) => includes(node.promptIDs, promptId)),
 );
 
-const makeNetworkNodesForPrompt = () => getNetworkNodesForPrompt;
-
 /**
  * makeNetworkNodesForOtherPrompts()
  *
@@ -186,6 +155,3 @@ export const getNetworkNodesForOtherPrompts = createSelector(
   getNetworkNodesForType, getPromptId,
   (nodes, promptId) => filter(nodes, (node) => !includes(node.promptIDs, promptId)),
 );
-
-const makeNetworkNodesForOtherPrompts = () =>
-  getNetworkNodesForOtherPrompts;
