+++ lib/interviewer/selectors/name-generator.js
@@ -21,7 +21,6 @@ const defaultPanelConfiguration = {
 // MemoedSelectors
 
 const stageCardOptions = (_, props) => props.stage.cardOptions;
-const stageSortOptions = (_, props) => props.stage.sortOptions;
 const propPanels = (_, props) => props.stage.panels;
 
 const getIDs = createSelector(
@@ -47,8 +46,6 @@ export const getPromptModelData = createSelector(
   },
 )
 
-const makeGetPromptNodeModelData = () => getPromptModelData;
-
 // Returns any additional properties to be displayed on cards.
 // Returns an empty array if no additional properties are specified in the protocol.
 export const getCardAdditionalProperties = createSelector(
@@ -56,16 +53,6 @@ export const getCardAdditionalProperties = createSelector(
   (cardOptions) => (has(cardOptions, 'additionalProperties') ? cardOptions.additionalProperties : []),
 );
 
-// Returns the properties that are specified as sortable in sortOptions
-const getSortableFields = createSelector(
-  stageSortOptions,
-  (sortOptions) => (has(sortOptions, 'sortableProperties') ? sortOptions.sortableProperties : []),
-);
-
-const getInitialSortOrder = createSelector(
-  stageSortOptions,
-  (sortOptions) => (has(sortOptions, 'sortOrder') ? sortOptions.sortOrder : []),
-);
 
 export const getNodeIconName = createSelector(
   getProtocolCodebook,
@@ -75,8 +62,6 @@ export const getNodeIconName = createSelector(
     return (nodeInfo && nodeInfo[nodeType] && nodeInfo[nodeType].iconVariant) || 'add-a-person';
   },
 );
-
-const makeGetNodeIconName = () => getNodeIconName;
 
 export const makeGetPanelConfiguration = () => createSelector(
   propPanels,
