+++ lib/interviewer/selectors/prop.js
@@ -41,8 +41,6 @@ export const getSubjectType = createSelector(getStageSubject, (subject) => {
   return subject?.type ?? null;
 });
 
-const makeGetSubject = () => getStageSubject;
-
 export const getAdditionalAttributesSelector = createSelector(
   propStage,
   propPrompt,
@@ -56,9 +54,6 @@ export const getPromptVariable = createSelector(
   propPrompt,
   (prompt) => prompt?.variable ?? null,
 );
-
-// TODO: Not sure this needs to be a createSelector
-const makeGetPromptVariable = () => getPromptVariable;
 
 export const getPromptOtherVariable = createSelector(propPrompt, (prompt) => [
   prompt?.otherVariable ?? null,
