+++ lib/interviewer/selectors/protocol.js
@@ -1,11 +1,4 @@
 import { v4 as uuid } from 'uuid';
-import {
-  orderBy,
-  values,
-  mapValues,
-  omit,
-} from 'lodash';
-import { entityAttributesProperty } from '@codaco/shared-consts';
 import { get } from '../utils/lodash-replacements';
 import { createSelector } from '@reduxjs/toolkit';
 import { getStageSubject } from './prop';
@@ -21,69 +14,12 @@ const getActiveSession = (state) => (
   state.activeSessionId && state.sessions[state.activeSessionId]
 );
 
-const getLastActiveSession = (state) => {
-  if (Object.keys(state.sessions).length === 0) {
-    return {};
-  }
-
-  const sessionsCollection = values(mapValues(state.sessions, (session, sessionUUID) => ({
-    sessionUUID,
-    ...session,
-  })));
-
-  const lastActive = orderBy(sessionsCollection, ['updatedAt', 'caseId'], ['desc', 'asc'])[0];
-  return {
-    sessionUUID: lastActive.sessionUUID,
-    [entityAttributesProperty]: {
-      ...omit(lastActive, 'sessionUUID'),
-    },
-  };
-};
-
 const getInstalledProtocols = (state) => state.installedProtocols;
 
 const getCurrentSessionProtocol = createSelector(
   getActiveSession,
   getInstalledProtocols,
   (session, protocols) => protocols[session?.protocolUID],
-);
-
-// Use the protocol associated with the last active session, unless there is a protocol with
-// an `installationDate` that is more recent.
-const getLastActiveProtocol = (state) => {
-  const installedProtocols = getInstalledProtocols(state);
-
-  if (Object.keys(installedProtocols).length === 0) {
-    return null;
-  }
-
-  const lastActiveSession = getLastActiveSession(state);
-  const lastActiveAttributes = lastActiveSession[entityAttributesProperty];
-
-  const protocolsCollection = values(mapValues(installedProtocols, (protocol, protocolUID) => ({
-    protocolUID,
-    ...protocol,
-  })));
-
-  const lastInstalledProtocol = orderBy(protocolsCollection, ['installationDate'], ['desc'])[0];
-
-  if (
-    lastActiveAttributes
-    && lastActiveAttributes.updatedAt // Last active session exists
-    && lastActiveAttributes.updatedAt > lastInstalledProtocol.installationDate
-  ) {
-    return {
-      ...installedProtocols[lastActiveSession[entityAttributesProperty].protocolUID],
-      protocolUID: lastActiveSession[entityAttributesProperty].protocolUID,
-    };
-  }
-
-  return lastInstalledProtocol;
-};
-
-const getActiveProtocolName = createSelector(
-  getCurrentSessionProtocol,
-  (protocol) => protocol && protocol.name,
 );
 
 export const getAssetManifest = createSelector(
@@ -168,6 +104,3 @@ export const getCodebookVariablesForType = createSelector(
       ? codebook[subject.entity][subject.type].variables
       : codebook.ego.variables),
 );
-
-const makeGetCodebookVariablesForType = () =>
-  getCodebookVariablesForType;
