+++ lib/interviewer/selectors/utils.js
@@ -6,11 +6,3 @@ export const createDeepEqualSelector = createSelectorCreator(
   defaultMemoize,
   isEqual,
 );
-
-// If condition is not met, throws. When thrown from a selector, the error
-// will be handled by stage boundary and displayed to the user.
-const assert = (condition, errorMessage) => {
-  if (!condition) {
-    throw new Error(errorMessage);
-  }
-};
