+++ lib/interviewer/utils/Validations.js
@@ -23,20 +23,6 @@ const coerceArray = (value) => {
   return [];
 };
 
-const validateUrl = (message) => (url) => {
-  try {
-    const constructURL = new URL(url);
-
-    if (constructURL.protocol !== 'http:' && constructURL.protocol !== 'https:') {
-      throw new Error('Invalid protocol');
-    }
-  } catch (e) {
-    return message || 'Please enter a valid URL, including http:// or https://';
-  }
-
-  return undefined;
-};
-
 const required = (message) => (value) => {
   const isEmptyString = isString(value) && value.length === 0;
