+++ lib/interviewer/utils/lodash-replacements.js
@@ -1,26 +1,3 @@
-/* eslint-disable import/prefer-default-export */
-// import { get as lodashGet } from 'lodash';
-
-/**
- * Helper function for verifying that a replacement is functioning the same as the original
- *
- * Usage:
- * export verify(myGet, lodashGet);
- *
- */
-// eslint-disable-next-line no-unused-vars
-const verify = (fn, otherFn) => (...args) => {
-  const result = fn(...args);
-  const otherResult = otherFn(...args);
-  if (result !== otherResult) {
-    // eslint-disable-next-line no-console
-    console.log('return values do not match', result, otherResult);
-    // eslint-disable-next-line no-debugger
-    debugger;
-  }
-  return result;
-};
-
 const pathReducer = (acc, part) => {
   // If part is an array, call pathReducer on each element
   if (Array.isArray(part)) {
