+++ lib/network-exporters/formatters/graphml/helpers.js
@@ -1,5 +1,4 @@
 import { isNil } from 'lodash';
-import { VariableType } from '@codaco/shared-consts';
 import { entityAttributesProperty } from '@codaco/shared-consts';
 
 const getEntityAttributes = (node) =>
@@ -18,8 +17,6 @@ const formatXml = (xml, tab = '\t') => {
   });
   return formatted.substring(1, formatted.length - 3);
 };
-
-const VariableTypeValues = Object.freeze(Object.values(VariableType));
 
 /**
  * For a given key, return a valid Graphml data 'type' for encoding
@@ -78,21 +75,6 @@ const getVariableInfo = (codebook, type, entity, key) =>
 const getEgoVariableInfo = (codebook, key) =>
   codebook.ego && codebook.ego.variables && codebook.ego.variables[key];
 
-/**
- * Determine if a given variable is one of the valid NC vattribute types
- * @param {*} codebook
- * @param {*} type
- * @param {*} element
- * @param {*} key
- */
-const codebookExists = (codebook, type, element, key) => {
-  const variableInfo = getVariableInfo(codebook, type, element, key);
-  return (
-    variableInfo &&
-    variableInfo.type &&
-    VariableTypeValues.includes(variableInfo.type)
-  );
-};
 
 /**
  * Get the 'type' of a given variable from the codebook
