+++ lib/network-exporters/utils/general.js
@@ -29,11 +29,6 @@ export const getEntityAttributes = (entity) =>
 
 const escapeFilePart = (part) => part.replace(/\W/g, '');
 
-const sleep =
-  (time = 2000) =>
-  (passThrough) =>
-    new Promise((resolve) => setTimeout(() => resolve(passThrough), time));
-
 export const makeFilename = (prefix, entityType, exportFormat, extension) => {
   let name = prefix;
   if (extension !== `.${exportFormat}`) {
@@ -77,7 +72,6 @@ export const getFilePrefix = (session, protocol, unifyNetworks) => {
     return sanitizeFilename(protocol.name);
   }
 
-  return `${sanitizeFilename(session.sessionVariables[caseProperty])}_${
-    session.sessionVariables[sessionProperty]
-  }`;
+  return `${sanitizeFilename(session.sessionVariables[caseProperty])}_${session.sessionVariables[sessionProperty]
+    }`;
 };
