+++ lib/network-query/rules.js
@@ -5,7 +5,6 @@ import {
 import { operators } from './predicate';
 import predicate from './predicate';
 
-
 const singleEdgeRule =
   ({ type, attribute, operator, value: other }) =>
     (node, edges) => {
@@ -205,7 +204,7 @@ const createRule = (type, options, f) => {
  * const result = rule(node, edges); // returns boolean
  * ```
  */
-const getRule = ({ type, options }) => {
+export const getRule = ({ type, options }) => {
   switch (type) {
     case 'alter':
       return createRule('alter', options, nodeRule);
@@ -219,7 +218,7 @@ const getRule = ({ type, options }) => {
 };
 
 // As above, but for rules matching single array or edge
-const getSingleRule = ({ type, options }) => {
+export const getSingleRule = ({ type, options }) => {
   switch (type) {
     case 'alter':
       return createRule('alter', options, singleNodeRule);
@@ -231,10 +230,5 @@ const getSingleRule = ({ type, options }) => {
       return () => false;
   }
 };
-
-const singleRule = getSingleRule;
-const alter = nodeRule;
-const edge = edgeRule;
-const ego = egoRule;
