+++ lib/ui/components/Fields/DatePicker/helpers.js
@@ -7,8 +7,6 @@ export const isEmpty = (value) => value === null || value === '';
 
 export const getFirstDayOfMonth = (dateObj) => DateTime.fromObject({ ...dateObj, day: 1 }).toFormat('c');
 
-const asNullObject = (keys) => keys.reduce((acc, key) => ({ ...acc, [key]: null }), {});
-
 const getProperties = (obj) => Object.keys(obj)
   .reduce((acc, key) => {
     if (!obj[key]) { return acc; }
