+++ lib/ui/utils/CSSVariables.js
@@ -22,17 +22,3 @@ export const getCSSVariableAsString = (variableName) => CSSVariable(variableName
 export const getCSSVariableAsNumber = (variableName) => parseInt(CSSVariable(variableName), 10);
 
 export const getCSSVariableAsObject = (variableName) => JSON.parse(CSSVariable(variableName));
-
-const getCSSVariable = (variableName) => {
-  const variable = CSSVariable(variableName);
-
-  try {
-    return JSON.parse(variable);
-  } catch (e) {
-    if (parseInt(variable, 10).toString() === variable) {
-      return parseInt(variable, 10);
-    }
-
-    return variable;
-  }
-};
