+++ lib/ui/utils/config.js
@@ -13,12 +13,6 @@ const ALLOWED_MARKDOWN_TAGS = [
   'hr',
 ];
 
-const ALLOWED_MARKDOWN_PROMPT_TAGS = [
-  'p',
-  'em',
-  'strong',
-];
-
 const ALLOWED_MARKDOWN_LABEL_TAGS = [
   'br',
   'p',
