+++ next.config.mjs
@@ -22,7 +22,7 @@ const config = {
   },
   webpack: (config) => {
     config.module.rules.push({
-      test: /\.(jpe?g|png|svg|gif|ico|eot|ttf|woff|woff2|mp4|pdf|webm|txt|wav)$/,
+      test: /\.(jpe?g|png|svg|gif|ico|eot|ttf|woff|woff2|mp4|pdf|webm|txt|mp3)$/,
       type: 'asset/resource',
       generator: {
         filename: 'static/chunks/[path][name].[hash][ext]'
